pyproject.toml
[tool.poetry]
name = "orm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "b"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
isort = "^5.12.0"
black = "^23.7.0"

[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
